设置entry 当前路径

+++ include/xrun/utils/string.h
@@ -10,8 +10,8 @@
 typedef struct xr_string_s xr_string_t;
 
 struct xr_string_s {
-  int length;
-  int capacity;
+  size_t length;
+  size_t capacity;
   char *string;
 };
